feat: Implement welcome screen and fade-in animation
Adds a welcome screen to the application that is displayed on initial load. Includes a CSS animation for a fade-in effect on elements. The welcome screen can be dismissed by clicking a "Continue" button, transitioning the user to the main input form. The `handleReset` function is updated to re-show the welcome screen.